Move resolveNetworkAddress to a separate file
